+++ driver/cl_options.h
@@ -143,6 +143,9 @@ extern cl::opt<unsigned> fWarnStackSize;
 #if LDC_LLVM_SUPPORTED_TARGET_SPIRV || LDC_LLVM_SUPPORTED_TARGET_NVPTX
 extern cl::list<std::string> dcomputeTargets;
 extern cl::opt<std::string> dcomputeFilePrefix;
+extern cl::opt<bool> fembedDCompute;
+#else
+constexpr bool fembedDCompute = false;
 #endif
 
 #if defined(LDC_DYNAMIC_COMPILE)